+++ include/ccl_power.h
@@ -34,11 +34,12 @@ double ccl_nonlin_matter_power(ccl_cosmology * cosmo, double k, double a,int * s
  * Compute the linear power spectrum and create a 2d spline P(k,z) to be stored
  * in the cosmology structure.
  * @param cosmo Cosmological parameters
+ * param psp The linear power spectrum to use for Boltzmann solvers, if any.
  * @param status Status flag. 0 if there are no errors, nonzero otherwise.
  * For specific cases see documentation for ccl_error.c
  * @return void
  */
-void ccl_cosmology_compute_linear_power(ccl_cosmology * cosmo, int* status);
+void ccl_cosmology_compute_linear_power(ccl_cosmology * cosmo, ccl_f2d_t *psp, int* status);
 
 /**
  * Compute the non-linear power spectrum and create a 2d spline P(k,z) to be stored