+++ src/ccl_power.c
@@ -7,8 +7,6 @@
 #include <gsl/gsl_spline.h>
 #include <gsl/gsl_errno.h>
 
-#include <class.h> /* from extern/ */
-
 #include "ccl.h"
 #include "ccl_emu17.h"
 #include "ccl_emu17_params.h"
@@ -432,7 +430,7 @@ static void ccl_cosmology_spline_nonlinpower(
 INPUT: ccl_cosmology * cosmo
 TASK: compute linear power spectrum
 */
-void ccl_cosmology_compute_linear_power(ccl_cosmology* cosmo, int* status) {
+void ccl_cosmology_compute_linear_power(ccl_cosmology* cosmo, ccl_f2d_t *psp, int* status) {
   if ((cosmo->config.transfer_function_method != ccl_boltzmann_class &&
        cosmo->config.transfer_function_method != ccl_transfer_none) &&
       (fabs(cosmo->params.mu_0) > 1e-14 || fabs(cosmo->params.sigma_0) > 1e-14)) {
@@ -466,7 +464,7 @@ void ccl_cosmology_compute_linear_power(ccl_cosmology* cosmo, int* status) {
         break;
 
       case ccl_boltzmann_class:
-        ccl_cosmology_compute_linpower_class(cosmo, status);
+        ccl_cosmology_spline_linpower_musigma(cosmo, psp, status);
         break;
 
       default: {